Take device screenshot
When the ‘Take screenshot’ action is invoked it will take a screenshot of the selected device and save it to a file. If there is more than one device connected, a dialog will be displayed where a device can be selected. If there is only one device connected, no dialog will be presented and the only device available is used.